requirements.txt
ldap3
argparse
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
